refactor: :recycle: Changed behavior of AsyncCreatableSelect for better responsiveness
