+++ module/addHorseForm/components/form/components/fatherSideInformation/FatherInformation.tsx
@@ -1,13 +1,6 @@
 import React from 'react';
 import { FormTemplateDetails } from '../../../template/FormTemplateDetails';
 
-const options = {
-  parent: [
-    { value: 'klacz1', label: 'klacz1' },
-    { value: 'klacz3', label: 'klacz3' },
-  ],
-};
-
 export const FatherSideInformation = () => {
-  return <FormTemplateDetails familySide="father" options={options} />;
+  return <FormTemplateDetails familySide="father" gender="stallion" label="add father" />;
 };
