+++ module/addHorseForm/components/form/utils/helpers.tsx
@@ -24,19 +24,6 @@ export const steps = [
   { name: 'Summary', component: <Summary /> },
 ] as const;
 
-export const readFileAsDataURL = (file: File): Promise<string> => {
-  return new Promise((resolve, reject) => {
-    const reader = new FileReader();
-    reader.onload = (event) => {
-      resolve(event.target?.result as string);
-    };
-    reader.onerror = (error) => {
-      reject(error);
-    };
-    reader.readAsDataURL(file);
-  });
-};
-
 export const formFields = {
   input: [
     { name: 'name', placeholder: 'name' },
