+++ module/addHorseForm/components/template/FormTemplateDetails.tsx
@@ -1,27 +1,33 @@
 import React from 'react';
 import { useFormContext } from 'react-hook-form';
-import { useAsyncControllerSelect } from '@/shared/asyncControllerSelect/useAsyncControllerSelect';
 import { FormControl, Grid, GridItem } from '@chakra-ui/react';
-import { ParentSideInfo } from '../../utils/types';
-import ParentField from './ParentsField';
+import { GenderType } from '@/utils/types';
+import { AyncCreateableSelect } from '@/shared/ayncCreateableSelect/AyncCreateableSelect';
 
 interface FormTemplateDetailsProps {
   familySide: 'father' | 'mother';
-  options: ParentSideInfo;
   placeholder?: string;
+  gender: GenderType;
+  label: string;
 }
 
-export const FormTemplateDetails = ({ familySide, options, placeholder }: FormTemplateDetailsProps) => {
+export const FormTemplateDetails = ({ familySide, gender, label, placeholder }: FormTemplateDetailsProps) => {
   const {
+    control,
     formState: { errors },
   } = useFormContext();
-  const { familyMemberOptions } = useAsyncControllerSelect(options);
-  console.log(errors);
+
   return (
     <FormControl isInvalid={Boolean(errors[familySide])}>
       <Grid templateColumns="repeat(2, 1fr)" gap={6}>
         <GridItem colSpan={1}>
-          <ParentField familySide={familySide} familyMemberOptions={familyMemberOptions} placeholder={placeholder} />
+          <AyncCreateableSelect
+            control={control}
+            gender={gender}
+            name={familySide}
+            label={label}
+            placeholder={placeholder}
+          />
         </GridItem>
       </Grid>
     </FormControl>
