+++ module/profile/components/manageHorseAction/components/horseEditData/editForm/utils/helpers.ts
@@ -1,5 +1,3 @@
-import { OptionType } from '@/utils/types';
-
 export const formFields = {
   input: [
     { name: 'name', placeholder: 'name', label: 'Change name' },
@@ -12,11 +10,3 @@ export const formFields = {
     { value: 'stallion', content: 'stallion' },
   ],
 };
-
-export const filterhorses = (inputValue: string, arr: OptionType[]) => {
-  return arr.filter((i) => i.label.toLowerCase().includes(inputValue.toLowerCase()));
-};
-export const createOption = (label: string) => ({
-  label,
-  value: label,
-});
